refactor: fix service structure
